don't insist on binaries in ${GOPATH}/bin
${GOPATH} is a pathspec and can be a list of directories, not only a
single one. For example, this is a valid and not uncommon value:

    /home/mark/go-theirs:/home/mark/go-mine:/home/mark/go-ours

(Above pathspec will install third-party libraries in its own tree,
retaining any customized versions in separate locations.)

And then, 'godep', 'upx', 'golint', and friends could have been installed
globally, for all users, not in ${GOPATH}/bin. It is not necessary to insist
on a specific location as long as those binaries are found by BASH.