aci: sort contents, and pull manifest to the front
If the manifest is the first file then tools do not need to download
the whole image file in order to detect changes or updates.

Sorted contents of ACIs allow for easier comparison, and usage of
tools such as zsync, and deduplication on the server. The price,
sorting by 'tar', is cheap.

Parameter 'f' for 'tar' must be followed by the filename for recent
versions of 'tar'. Mind the order!

closes #210