CI: install curl 7.36.0 if we encounter an older version
Travis treats us with Ubuntu 14.04, which comes with curl 7.35, which
in turn is very old and long obsoleted.

The PPA referenced here does offer no more than ettercap, not used here,
and curl 7.36. Our minimum requirement. In case of an upgrade to
Ubuntu 16.04 that version gets ignored: Xenial comes with 7.47.0.