aci-builder: pin the build-date, if given ex tunc
If the user has provided a build-date (as aci::annotation), then don't
overwrite it. Instead, use it as ceiling to modification datetime of files.

If everything else has been created in a deterministic manner, this
results in completely reproducible builds.

Please note that you can use templating to set this value.