[aci-tester] switch to maintained bats-core v1.1.0
The bats project has moved to bats-core and has recieved several
backwards-compatible fixes in its newer versions. These updates may fix
some of the issues with hanging tests and other bad or otherwise flaky
behavior.

As a prerequisite, I also had to update a bunch of other stuff (mostly
the example ACIs) in order for the build to complete successfully.